adjust webspace configuration before deployment
