+++ web/.htaccess
@@ -5,8 +5,6 @@
 # to each configured DirectoryIndex file (e.g. index.php, index.html, index.pl).
 DirectoryIndex website.php
 
-SetEnv APP_ENV prod
-
 <IfModule mod_rewrite.c>
     RewriteEngine On
